j10

+++ srcs/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: jacens <jacens@student.le-101.fr>          +:+   +:    +:    +:+     */
 /*                                                 #+#   #+    #+    #+#      */
 /*   Created: 2019/10/10 10:59:20 by jacens       #+#   ##    ##    #+#       */
-/*   Updated: 2019/10/14 14:59:35 by jacens      ###    #+. /#+    ###.fr     */
+/*   Updated: 2019/10/16 16:40:35 by jacens      ###    #+. /#+    ###.fr     */
 /*                                                         /                  */
 /*                                                        /                   */
 /* ************************************************************************** */
@@ -15,7 +15,7 @@
 
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
-	unsigned int	i;
+	size_t			i;
 	char			*ptr;
 
 	i = 0;
